refactor: moved to feature-based structure
Moved to a feature-based, instead of type-based structure (e.g. `ToDos` feature instead of organizing by services).
