+++ src/NoPlan.Api/Endpoints/V1/ToDos/GetToDoEndpoint.cs
@@ -1,4 +1,4 @@
-﻿using NoPlan.Api.Services;
+﻿using NoPlan.Api.Features.ToDos;
 using NoPlan.Contracts.Requests.V1.ToDos;
 using NoPlan.Contracts.Responses.V1.ToDos;
 using NoPlan.Infrastructure.Data.Models;
